TAN-227: extract storage maintenance from main.rs to satisfy the file-size gate
The Lint Backend gate caps touched .rs files at 2000 lines; adding the
smoke subcommand tripped it on engine/src/main.rs (2849 lines, almost all
pre-existing). Moves the self-contained storage doctor/cleanup machinery
(report types, context-run archival, jsonl sharding, report printers) into
engine/src/storage_maintenance.rs. No behavior change: smoke run green,
storage doctor exercised, docs parity green; main.rs now 1803 lines.
